toolkit/content/tests/widgets/test_moz_input_tel.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>MozInputTel Tests</title>
    <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
    <link rel="stylesheet" href="chrome://global/skin/in-content/common.css" />
    <link
      rel="stylesheet"
      href="chrome://mochikit/content/tests/SimpleTest/test.css"
    />
    <script
      type="module"
      src="chrome://global/content/elements/moz-input-tel.mjs"
    ></script>
    <script src="lit-test-helpers.js"></script>
    <script class="testbody" type="application/javascript">
      let testHelpers = new InputTestHelpers();
      let html;

      add_setup(async function setup() {
        ({ html } = await testHelpers.setupLit());
        testHelpers.setupTests({
          templateFn: (attrs, children) =>
            html`<moz-input-tel ${attrs}>${children}</moz-input-tel>`,
        });
      });

      add_task(async function testMozInputTelProperties() {
        await testHelpers.testCommonInputProperties("moz-input-tel", {
          type: "tel",
        });
      });

      add_task(async function testMozInputTelEvents() {
        await testHelpers.testTextBasedInputEvents("moz-input-tel");
      });
    </script>
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>
</html>
